pyproject.toml
[project]
name = "rag-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu==1.8.0",
    "langchain==0.3.0",
    "langchain-community==0.3.0",
    "langchain-openai==0.2.0",
    "pypdf==4.3.1",
    "python-dotenv==1.0.1",
    "streamlit==1.38.0",
    "tiktoken==0.7.0",
]
